Add tasks as separate classes
