+++ src/lesson_1/TaskTwo.java
@@ -1,0 +1,14 @@
+package lesson_1;
+
+public class TaskTwo {
+
+    public static void main(String[] args) {
+
+        sayHelloMe("Evgeniy");
+    }
+
+    public static void sayHelloMe (String name) {
+        System.out.println("Hello, " + name);
+    }
+
+}
